+++ client/Router/index.js
@@ -1,0 +1,11 @@
+export {
+  createClientResolver,
+  createClientRouter,
+  createResolver,
+  historyMiddlewares,
+  render,
+  default,
+} from './Router'
+
+export { default as routeConfig } from './routeConfig'
+export { default as paths } from './paths'
