+++ client/Router/paths.js
@@ -1,0 +1,3 @@
+export default {
+  unauthorized: '/unauthorized',
+}
